pr feedback

+++ craft_archives/repo/apt_key_manager.py
@@ -257,7 +257,7 @@ class AptKeyManager:
         # If the keyring exists but does not contain the key, remove it and
         # install a fresh one.
         keyring_path = get_keyring_path(key_id, base_path=self._keyrings_path)
-        with contextlib.suppress(NotADirectoryError):
+        if keyring_path.parent.is_dir():
             keyring_path.unlink(missing_ok=True)
 
         key_path = self.find_asset_with_key_id(key_id=key_id)
